+++ frontend/src/styles/Footer.module.css
@@ -1,26 +1,10 @@
-.footer {
-  height: 70px;
-  margin-top: auto;
-  text-align: center;
-  width: 100%;
-  bottom: 0;
-  padding-top: 10px;
-  position: absolute;
-  left: 0;
-  bottom: 0;
-  font-size: 14px;
-}
-
-.footer a {
-  transition: all 0.2s ease-in-out;
-}
-
-.footer a:hover {
-  color: #4b92db;
-}
-
 .social {
   color: #636363;
   font-size: 14px;
   margin: 8px 8px 0;
+  transition: all 0.2s ease-in-out;
+}
+
+.social:hover {
+  color: #4b92db;
 }
